+++ common/src/main/java/net/mca/item/ItemsMCA.java
@@ -121,6 +121,32 @@ public interface ItemsMCA {
             .addPage(new CenteredTextPage("cult_0", 3))
             .addPage(new ScribbleTextPage(MCA.locate("textures/gui/scribbles/goat.png"), ""))));
 
+    RegistrySupplier<Item> BOOK_CULT_1 = register("book_cult_1", () -> new ExtendedWrittenBookItem(baseProps(), new Book("cult_1")
+            .setBackground(MCA.locate("textures/gui/books/cult.png"))
+            .setTextFormatting(Formatting.DARK_RED)
+            .addPage(new TitlePage("cult_1", Formatting.DARK_RED))
+            .addPage(new CenteredTextPage("cult_1", 0))
+            .addPage(new CenteredTextPage("cult_1", 1))
+            .addPage(new CenteredTextPage("cult_1", 2))
+            .addPage(new CenteredTextPage("cult_1", 3))
+            .addPage(new ScribbleTextPage(MCA.locate("textures/gui/scribbles/goat.png"), ""))));
+
+    RegistrySupplier<Item> BOOK_CULT_2 = register("book_cult_2", () -> new ExtendedWrittenBookItem(baseProps(), new Book("cult_2")
+            .setBackground(MCA.locate("textures/gui/books/cult.png"))
+            .setTextFormatting(Formatting.DARK_RED)
+            .addPage(new TitlePage("cult_2", Formatting.DARK_RED))
+            .addPage(new CenteredTextPage("cult_2", 0))
+            .addPage(new CenteredTextPage("cult_2", 1))
+            .addPage(new CenteredTextPage("cult_2", 2))
+            .addPage(new CenteredTextPage("cult_2", 3))
+            .addPage(new ScribbleTextPage(MCA.locate("textures/gui/scribbles/goat.png"), ""))));
+
+    RegistrySupplier<Item> BOOK_CULT_ANCIENT = register("book_cult_ancient", () -> new ExtendedWrittenBookItem(baseProps(), new Book("cult_ancient")
+            .setBackground(MCA.locate("textures/gui/books/cult.png"))
+            .setTextFormatting(Formatting.DARK_RED)
+            .addPage(new TitlePage("cult_ancient", Formatting.DARK_RED))
+            .addPage(new CenteredTextPage("cult_ancient", 0))));
+
     RegistrySupplier<Item> LETTER = register("letter", () -> new ExtendedWrittenBookItem(baseProps().maxCount(1), new Book("letter", null)
             .setBackground(MCA.locate("textures/gui/books/paper.png"))));